+++ pkgs/test/test/runner/compiler_runtime_matrix_test.dart
@@ -102,7 +102,7 @@ void main() {
         if (runtime.isDartVM) {
           test('forwards stdout/stderr', () async {
             await d.file('test.dart', _testWithStdOutAndErr).create();
-            var test = await runTest(testArgs);
+            var test = await runTest(testArgs, reporter: 'silent');
 
             expect(test.stdout, emitsThrough('hello'));
             expect(test.stderr, emits('world'));